+++ cvat-ui/src/containers/header/header.tsx
@@ -7,7 +7,7 @@ import { connect } from 'react-redux';
 import getCore from 'cvat-core-wrapper';
 import HeaderComponent from 'components/header/header';
 import { SupportedPlugins, CombinedState } from 'reducers/interfaces';
-import { logoutAsync } from 'actions/auth-actions';
+import { logoutAsync, authActions } from 'actions/auth-actions';
 import { switchSettingsDialog } from 'actions/settings-actions';
 
 const core = getCore();
@@ -25,20 +25,27 @@ interface StateToProps {
     uiVersion: string;
     switchSettingsShortcut: string;
     settingsDialogShown: boolean;
+    changePasswordDialogShown: boolean;
+    changePasswordFetching: boolean;
+    renderChangePasswordItem: boolean;
 }
 
 interface DispatchToProps {
     onLogout: () => void;
     switchSettingsDialog: (show: boolean) => void;
+    switchChangePasswordDialog: (show: boolean) => void;
 }
 
 function mapStateToProps(state: CombinedState): StateToProps {
     const {
         auth: {
             fetching: logoutFetching,
+            fetching: changePasswordFetching,
             user: {
                 username,
             },
+            showChangePasswordDialog: changePasswordDialogShown,
+            allowChangePassword: renderChangePasswordItem,
         },
         plugins: {
             list,
@@ -68,6 +75,9 @@ function mapStateToProps(state: CombinedState): StateToProps {
         uiVersion: packageVersion.ui,
         switchSettingsShortcut: normalizedKeyMap.SWITCH_SETTINGS,
         settingsDialogShown,
+        changePasswordFetching,
+        changePasswordDialogShown,
+        renderChangePasswordItem,
     };
 }
 
@@ -75,6 +85,9 @@ function mapDispatchToProps(dispatch: any): DispatchToProps {
     return {
         onLogout: (): void => dispatch(logoutAsync()),
         switchSettingsDialog: (show: boolean): void => dispatch(switchSettingsDialog(show)),
+        switchChangePasswordDialog: (show: boolean): void => (
+            dispatch(authActions.switchChangePasswordDialog(show))
+        ),
     };
 }